requirements/base.txt
django==1.10.4
django-model-utils==2.6
django-allauth==0.29.0
django-crispy-forms==1.6.1
django-taggit==0.21.3
django-contrib-comments==1.7.3
django-mptt==0.8.6
django-autoslug==1.9.3
django-imagekit==3.3
django-braces==1.10.0
django-notifications-hq==1.2
django-bootstrap-pagination==1.6.2
django-simple-captcha==0.5.3
django-pagedown==0.1.1
django-robots==3.0
django-anymail[sendgrid]==0.10
-e git+https://github.com/zmrenwu/django-activity-stream.git@master#egg=actstream

Markdown==2.6.7
Pygments==2.1.3
Pillow==3.4.2
python-slugify==1.2.1
bleach==1.5.0
timeago==1.0.7